+++ Biblioteca_InterfazSaldosDiarios/Models/Vencim.cs
@@ -12,7 +12,7 @@ namespace Biblioteca_InterfazSaldosDiarios.Models
         public string Agencia { get; set; }
         public string NumCuenta { get; set; }
         public string SufijoCuenta { get; set; }
-        public float Saldo { get; set; }
+        public decimal Saldo { get; set; }
         public DateTime FechaVen { get; set; }
         public int NumRegistros { get; set; }
         public float Intereses { get; set; }
